+++ plugins/actix-web/src/web.rs
@@ -774,4 +774,12 @@ impl<'a> ServiceConfig<'a> {
         self.inner.external_resource(name, url);
         self
     }
+
+    /// Proxy for [`actix_web::web::ServiceConfig::app_data`](https://docs.rs/actix-web/3.3.2/actix_web/web/struct.ServiceConfig.html#method.app_data).
+    ///
+    /// **NOTE:** This doesn't affect spec generation.
+    pub fn app_data<U: 'static>(&mut self, data: U) -> &mut Self {
+        self.inner.app_data(data);
+        self
+    }
 }