+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add support for actix-web-macros methods routing [PR#289](https://github.com/wafflespeanut/paperclip/pull/289)
 - Actix plugin: add an empty impl for actix-web `ReqData<T>`
 - Add support for the `#[serde(skip)]` attribute in structs and enums.
+- Added new method `trim_base_path` to trim the api base path from all method paths.
 
 ### Changed